added BaseRegistrator
